+++ Plugin/CleverTapUnity/iOS/CleverTapSDK.framework/PrivateHeaders/CTConstants.h
@@ -29,6 +29,11 @@ extern NSString *const kCTNotifViewedApiDomain;
 #define CLTAP_NOTIFICATION_VIEWED_EVENT_NAME @"Notification Viewed"
 #define CLTAP_GEOFENCE_ENTERED_EVENT_NAME @"Geocluster Entered"
 #define CLTAP_GEOFENCE_EXITED_EVENT_NAME @"Geocluster Exited"
+
+#define CLTAP_SIGNED_CALL_OUTGOING_EVENT_NAME @"SCOutgoing"
+#define CLTAP_SIGNED_CALL_INCOMING_EVENT_NAME @"SCIncoming"
+#define CLTAP_SIGNED_CALL_END_EVENT_NAME @"SCEnd"
+
 #define CLTAP_PREFS_LAST_DAILY_PUSHED_EVENTS_DATE @"lastDailyEventsPushedDate"
 #define CLTAP_SYSTEM_VERSION_LESS_THAN(v) ([[[UIDevice currentDevice] systemVersion] compare:v options:NSNumericSearch] == NSOrderedAscending)
 #define CLTAP_APP_LAUNCHED_EVENT @"App Launched"